requirements.txt
# Core
openai>=1.0.0
python-dotenv>=1.0.0
sympy>=1.12

# Backend API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Tools
wolframalpha>=5.0.0
tavily-python>=0.3.0

# OCR (Person 1 — choose one)
pytesseract>=0.3.10
Pillow>=10.0.0
# easyocr>=1.7.0  # Alternative: heavier but no Tesseract install needed

# Testing
pytest>=7.4.0
httpx>=0.27.0
